discord_bot/pyproject.toml
[tool.poetry]
name = "discord_bot"
version = "0.1.0"
description = ""
authors = ["Maxwell Farver"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.0"
requests = "^2.24.0"
python-dotenv = "^0.14.0"
pytest = "^6.1.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
